+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 
 - **Breaking Change**: in interactive mode or ``-c`` mode without shell-name, toplevel ``$0`` indicates ``argv[0]``
 - **Breaking Change**: unbalanced brace expansions are semantic error
+- **Breaking Change**: change token format error with semantic error
 - show stack trace of ignored exceptions within finally/defer block
 
 #### Builtin